requirements.txt
streamlit
pymongo
python-dotenv
sentence-transformers
tqdm
groq
fpdf
